factor out pagination slider logic
